requirements.txt
# ========================
# Core Framework
# ========================
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.1
python-multipart==0.0.9
sqlmodel==0.0.21
asyncpg==0.27.0

# ========================
# Environment & Config
# ========================
python-dotenv==1.0.1
pydantic-settings==2.9.1

# ========================
# Database (Supabase)
# ========================
supabase==2.8.1

# ========================
# AI/ML Dependencies
# ========================
openai==1.52.0
numpy>=1.24.0

# ========================
# HTTP Clients
# ========================
httpx==0.27.2
requests==2.32.3

# ========================
# Auth & Security
# ========================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.29.0
